Added Scene transition to talking scene
Added a script for loading the scene the player needs to be taken to talk to the other character. In the inspector the LoadNeededScene method should be called and requires the name of the character to be entered. The name should always be written out with capital letters at the start of the name with no spaces eg KieranClements