Random text box spawning and small adjustments
Created a system to have text boxes spawn on random parts of the screen, have an issue where the text of the text boxes is too small and doesn't scale with the boxes. Will need to figure out the issue later. Stopped the player from moving when talking to a NPC, before all that was stopped was the player camera and not the movement itself.