continued demo level development + bug fixes & adjustments
I continued the demo level development by adding the conversation partner and setting everything up for that. I also added a script to adjust to dialog box to match a phone size and a normal text box. Also adjusted some matters with the stress meter, such as it deleting itself if it detects another one, resetting variables once levels change and creating a gameobject that redirects all fungus conversation systems to the stress meter since i wouldn't be able to set the stress meter in the inspector due to the stress meter originating from another scene.